# Improve performance of String.ToUpper and friends