--- a/src/System.Private.CoreLib/shared/System/Text/Utf16Utility.cs
+++ b/src/System.Private.CoreLib/shared/System/Text/Utf16Utility.cs
@@ -13,7 +13,7 @@ namespace System.Text
         /// Returns true iff the UInt32 represents two ASCII UTF-16 characters in machine endianness.
         /// </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool AllCharsInUInt32AreAscii(uint value)
+        internal static bool AllCharsInUIntAreAscii(uint value)
         {
             return (value & ~0x007F007Fu) == 0;
         }
@@ -30,7 +30,7 @@ namespace System.Text
         internal static uint ConvertAllAsciiCharsInUIntToLowercase(uint value)
         {
             // ASSUMPTION: Caller has validated that input value is ASCII.
-            Debug.Assert(AllCharsInUInt32AreAscii(value));
+            Debug.Assert(AllCharsInUIntAreAscii(value));
 
             // the 0x80 bit of each word of 'lowerIndicator' will be set iff the word has value >= 'A'
             uint lowerIndicator = value + 0x00800080u - 0x00410041u;
@@ -59,7 +59,7 @@ namespace System.Text
         internal static uint ConvertAllAsciiCharsInUIntToUppercase(uint value)
         {
             // ASSUMPTION: Caller has validated that input value is ASCII.
-            Debug.Assert(AllCharsInUInt32AreAscii(value));
+            Debug.Assert(AllCharsInUIntAreAscii(value));
 
             // the 0x80 bit of each word of 'lowerIndicator' will be set iff the word has value >= 'a'
             uint lowerIndicator = value + 0x00800080u - 0x00610061u;
@@ -74,6 +74,56 @@ namespace System.Text
             uint mask = (combinedIndicator & 0x00800080u) >> 2;
 
             return value ^ mask; // bit flip lowercase letters [a-z] => [A-Z]
+        }
+
+        /// <summary>
+        /// Given a UInt32 that represents two ASCII UTF-16 characters, returns true iff
+        /// the input contains one or more lowercase ASCII characters.
+        /// </summary>
+        /// <remarks>
+        /// This is a branchless implementation.
+        /// </remarks>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static bool UIntContainsAnyLowercaseAsciiChar(uint value)
+        {
+            // ASSUMPTION: Caller has validated that input value is ASCII.
+            Debug.Assert(AllCharsInUIntAreAscii(value));
+
+            // the 0x80 bit of each word of 'lowerIndicator' will be set iff the word has value >= 'a'
+            uint lowerIndicator = value + 0x00800080u - 0x00610061u;
+
+            // the 0x80 bit of each word of 'upperIndicator' will be set iff the word has value <= 'z'
+            uint upperIndicator = value + 0x00800080u - 0x007B007Bu;
+
+            // the 0x80 bit of each word of 'combinedIndicator' will be set iff the word has value >= 'a' and <= 'z'
+            uint combinedIndicator = (lowerIndicator ^ upperIndicator);
+
+            return (combinedIndicator & 0x00800080u) != 0;
+        }
+
+        /// <summary>
+        /// Given a UInt32 that represents two ASCII UTF-16 characters, returns true iff
+        /// the input contains one or more uppercase ASCII characters.
+        /// </summary>
+        /// <remarks>
+        /// This is a branchless implementation.
+        /// </remarks>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal static bool UIntContainsAnyUppercaseAsciiChar(uint value)
+        {
+            // ASSUMPTION: Caller has validated that input value is ASCII.
+            Debug.Assert(AllCharsInUIntAreAscii(value));
+
+            // the 0x80 bit of each word of 'lowerIndicator' will be set iff the word has value >= 'A'
+            uint lowerIndicator = value + 0x00800080u - 0x00410041u;
+
+            // the 0x80 bit of each word of 'upperIndicator' will be set iff the word has value <= 'Z'
+            uint upperIndicator = value + 0x00800080u - 0x005B005Bu;
+
+            // the 0x80 bit of each word of 'combinedIndicator' will be set iff the word has value >= 'A' and <= 'Z'
+            uint combinedIndicator = (lowerIndicator ^ upperIndicator);
+
+            return (combinedIndicator & 0x00800080u) != 0;
         }
     }
 }