This improves the performance of `String.ToUpper` / `ToUpperInvariant` / `ToLower` / `ToLowerInvariant`. This is a follow-up to https://github.com/dotnet/coreclr/pull/20275, which improved the performance of the corresponding `ReadOnlySpan<char>`-based extension methods.

```cs
[Benchmark]
[Arguments("")] // returns String.Empty (non-allocating)
[Arguments("ASCII")] // returns input argument (non-allocating)
[Arguments("MUCH LONGER ASCII MUCH LONGER ASCII")] // returns input argument (non-allocating)
[Arguments("This is a mix of UPPER and lower ASCII.")] // allocates
[Arguments("Транскрипто́мные техноло́гии")] // from ru.wikipedia.org front page, allocates, uses p/invoke
public string ToUpperInvariant(string argument)
{
    return argument.ToUpperInvariant();
}
```

|           Method |               Toolchain |             argument |       Mean |     Error |    StdDev | Scaled | ScaledSD |
|----------------- |------------------------ |--------------------- |-----------:|----------:|----------:|-------:|---------:|
| **ToUpperInvariant** | **3.0.0-preview1-27004-04** |                     (empty) |  **13.146 ns** | **0.8383 ns** | **0.8234 ns** |   **1.00** |     **0.00** |
| ToUpperInvariant |                   local |                      |   5.509 ns | 0.1373 ns | 0.1217 ns |   0.42 |     0.03 |
|                  |                         |                      |            |           |           |        |          |
| **ToUpperInvariant** | **3.0.0-preview1-27004-04** |                **ASCII** |  **20.057 ns** | **0.2416 ns** | **0.2142 ns** |   **1.00** |     **0.00** |
| ToUpperInvariant |                   local |                ASCII |   7.955 ns | 0.0687 ns | 0.0643 ns |   0.40 |     0.01 |
|                  |                         |                      |            |           |           |        |          |
| **ToUpperInvariant** | **3.0.0-preview1-27004-04** | **MUCH (...)ASCII [35]** |  **41.229 ns** | **0.5436 ns** | **0.4819 ns** |   **1.00** |     **0.00** |
| ToUpperInvariant |                   local | MUCH (...)ASCII [35] |  22.423 ns | 0.1994 ns | 0.1865 ns |   0.54 |     0.01 |
|                  |                         |                      |            |           |           |        |          |
| **ToUpperInvariant** | **3.0.0-preview1-27004-04** | **This (...)SCII. [39]** |  **97.540 ns** | **1.1472 ns** | **1.0731 ns** |   **1.00** |     **0.00** |
| ToUpperInvariant |                   local | This (...)SCII. [39] |  43.030 ns | 0.9679 ns | 1.5352 ns |   0.44 |     0.02 |
|                  |                         |                      |            |           |           |        |          |
| **ToUpperInvariant** | **3.0.0-preview1-27004-04** | **Транс(...)о́гии [28]** | **101.937 ns** | **1.3241 ns** | **1.1057 ns** |   **1.00** |     **0.00** |
| ToUpperInvariant |                   local | Транс(...)о́гии [28] |  92.030 ns | 1.0237 ns | 0.9075 ns |   0.90 |     0.01 |