@GrabYourPitchforks 2018-10-08
FYI I spot-checked corefx to ensure that there are existing tests for the `System.String` type. All tests passed on my local box.